code cleanup in views.py
