+++ labbackend/Views/refundandcancellation.py
@@ -12,6 +12,10 @@ from django.db.models import Q
 from django.core.mail import send_mail
 from django.views.decorators.csrf import csrf_exempt
 from django.conf import settings  #
+from datetime import datetime, date 
+from django.views.decorators.http import require_GET
+from django.views.decorators.http import require_http_methods
+
 @csrf_exempt
 def search_refund(request):
     if request.method == "GET":
@@ -396,3 +400,355 @@ def verify_and_process_cancellation(request):
             return JsonResponse({"error": str(e)}, status=500)
 
     return JsonResponse({"error": "Invalid request method."}, status=405)
+
+
+
+
+@csrf_exempt
+@require_http_methods(["GET"])
+def logs_api(request):
+    """Combined API endpoint for both refund and cancellation logs"""
+    try:
+        password = quote_plus('Smrft@2024')
+        client = MongoClient(
+                f'mongodb+srv://shinovalab:{password}@cluster0.xbq9c.mongodb.net/Lab?retryWrites=true&w=majority',
+                tls=True,
+                tlsCAFile=certifi.where()
+            )
+        db = client.Lab
+        patient_collection = db['labbackend_patient']
+        
+        # Get query parameters
+        log_type = request.GET.get('type', 'refund')  # Default to refund if not specified
+        start_date = request.GET.get('start_date')
+        end_date = request.GET.get('end_date')
+        
+        # Base query - default to current date if no dates provided
+        query = {}
+        
+        # Apply date filters
+        if start_date or end_date:
+            query['date'] = {}
+            if start_date:
+                start_date = datetime.strptime(start_date, '%Y-%m-%d')
+                query['date']['$gte'] = start_date
+            if end_date:
+                end_date = datetime.strptime(end_date, '%Y-%m-%d')
+                # Add 1 day to end_date to include the full day
+                end_date = end_date.replace(hour=23, minute=59, second=59)
+                query['date']['$lte'] = end_date
+        else:
+            # Default to current date if no dates provided
+            today = datetime.now()
+            today_start = today.replace(hour=0, minute=0, second=0, microsecond=0)
+            today_end = today.replace(hour=23, minute=59, second=59, microsecond=999999)
+            query['date'] = {'$gte': today_start, '$lte': today_end}
+        
+        # Additional optimization: Only fetch patients with refunded or cancelled tests
+        if log_type == 'refund':
+            # Add an additional filter to only fetch patients with refunded tests
+            # Assuming testname is stored as a string that we can do basic text matching on
+            query['testname'] = {'$regex': '"refund"\\s*:\\s*true', '$options': 'i'}
+        elif log_type == 'cancellation':
+            # Similar for cancellation
+            query['testname'] = {'$regex': '"cancellation"\\s*:\\s*true', '$options': 'i'}
+        
+        patients = list(patient_collection.find(query))
+        results = []
+        
+        if log_type == 'refund':
+            # Process refund logs
+            for patient in patients:
+                try:
+                    # Parse the testname JSON string
+                    tests = json.loads(patient.get('testname', '[]'))
+                    # Filter tests that have refund=true
+                    refundable_tests = [test for test in tests if test.get('refund') is True]
+                    
+                    # If there are refundable tests, add to results
+                    if refundable_tests:
+                        # Calculate total refund amount for this patient
+                        total_refund_amount = sum(float(test.get('amount', 0)) for test in refundable_tests)
+                        
+                        # List all refunded test names and their individual amounts
+                        refunded_test_details = [f"{test.get('testname', 'Unknown Test')} (₹{float(test.get('amount', 0)):.2f})" 
+                                               for test in refundable_tests]
+                        
+                        results.append({
+                            'id': str(patient.get('_id')),
+                            'patient_id': patient.get('patient_id'),
+                            'patientname': patient.get('patientname'),
+                            'bill_no': patient.get('bill_no'),
+                            'date': patient.get('date').isoformat() if isinstance(patient.get('date'), datetime) else str(patient.get('date')),
+                            'testname': ", ".join(refunded_test_details),
+                            'refund_amount': total_refund_amount,
+                            'refunded_tests': refundable_tests,  # Include full test objects for detailed info
+                            'refund_count': len(refundable_tests),  # Add count of refunded tests
+                            'reason': patient.get('refund_reason', 'Test Refunded')  # Try to get specific reason if available
+                        })
+                except (json.JSONDecodeError, AttributeError, KeyError) as e:
+                    # Skip if there's an error parsing the testname JSON
+                    print(f"Error processing patient {patient.get('_id')}: {str(e)}")
+                    continue
+        elif log_type == 'cancellation':
+            # Process cancellation logs
+            for patient in patients:
+                try:
+                    # Parse the testname JSON string
+                    tests = json.loads(patient.get('testname', '[]'))
+                    # Filter tests that are cancelled
+                    cancelled_tests = [test for test in tests if test.get('cancellation') is True]
+                    
+                    # If there are cancelled tests, add to results
+                    if cancelled_tests:
+                        # Calculate total cancelled amount
+                        total_cancelled_amount = sum(float(test.get('amount', 0)) for test in cancelled_tests)
+                        
+                        # List all cancelled test names with their individual amounts
+                        cancelled_test_details = [f"{test.get('testname', 'Unknown Test')} (₹{float(test.get('amount', 0)):.2f})" 
+                                               for test in cancelled_tests]
+                        
+                        results.append({
+                            'id': str(patient.get('_id')),
+                            'patient_id': patient.get('patient_id'),
+                            'patientname': patient.get('patientname'),
+                            'bill_no': patient.get('bill_no'),
+                            'date': patient.get('date').isoformat() if isinstance(patient.get('date'), datetime) else str(patient.get('date')),
+                            'testname': ", ".join(cancelled_test_details),
+                            'refund_amount': total_cancelled_amount,
+                            'cancelled_tests': cancelled_tests,  # Include full test objects for detailed info
+                            'cancel_count': len(cancelled_tests),  # Add count of cancelled tests
+                            'reason': patient.get('cancellation_reason', 'Test Cancelled')  # Try to get specific reason if available
+                        })
+                except (json.JSONDecodeError, AttributeError, KeyError) as e:
+                    # Skip if there's an error parsing the testname JSON
+                    print(f"Error processing patient {patient.get('_id')}: {str(e)}")
+                    continue
+        
+        return JsonResponse(results, safe=False)
+    except Exception as e:
+        print(f"Error in logs_api: {str(e)}")
+        return JsonResponse({"error": str(e)}, status=500)
+    
+
+
+
+@require_GET
+def dashboard_data(request):
+    try:
+        # Get date range and payment method from request parameters
+        from_date = request.GET.get('from_date')
+        to_date = request.GET.get('to_date')
+        payment_method = request.GET.get('payment_method')
+        # Set default to current date if no dates provided
+        if not from_date and not to_date:
+            today = date.today()
+            from_date = today.strftime('%Y-%m-%d')
+            to_date = today.strftime('%Y-%m-%d')
+        # Convert to datetime objects with timezone handling
+        if from_date:
+            from_date = datetime.strptime(from_date, '%Y-%m-%d')
+        if to_date:
+            # Set to end of day for inclusive filtering
+            to_date = datetime.strptime(to_date, '%Y-%m-%d').replace(hour=23, minute=59, second=59)
+        # MongoDB connection
+        password = quote_plus('Smrft@2024')
+        client = MongoClient(
+            f'mongodb+srv://shinovalab:{password}@cluster0.xbq9c.mongodb.net/Lab?retryWrites=true&w=majority',
+            tls=True,
+            tlsCAFile=certifi.where()
+        )
+        db = client.Lab
+        collection = db.labbackend_patient
+        # Build the query for date filtering
+        query = {}
+        if from_date and to_date:
+            query['date'] = {'$gte': from_date, '$lte': to_date}
+        elif from_date:
+            query['date'] = {'$gte': from_date}
+        elif to_date:
+            query['date'] = {'$lte': to_date}
+        # Add payment method filtering if specified
+        if payment_method:
+            if payment_method == "PartialPayment":
+                query['payment_method'] = {'$regex': 'PartialPayment', '$options': 'i'}
+            else:
+                # Check both direct payment_method and method inside PartialPayment
+                query['$or'] = [
+                    {'payment_method': {'$regex': f'"paymentmethod":"{payment_method}"', '$options': 'i'}},
+                    {'PartialPayment': {'$regex': f'"method":"{payment_method}"', '$options': 'i'}}
+                ]
+        # Execute the query to get filtered patients
+        patients = list(collection.find(query))
+        # Process the data for dashboard
+        total_patients = len(patients)
+        total_revenue = 0
+        # Payment method statistics
+        payment_methods = {
+            'Cash': 0,
+            'Card': 0,
+            'UPI': 0,
+            'Credit': 0,
+            'PartialPayment': 0
+        }
+        # Track revenue by payment method
+        payment_method_amounts = {
+            'Cash': 0,
+            'Card': 0,
+            'UPI': 0,
+            'Credit': 0,
+            'PartialPayment': 0
+        }
+        # Segment statistics
+        segments = {
+            'B2B': 0,
+            'Walk-in': 0,
+            'Home Collection': 0
+        }
+        # B2B client statistics
+        b2b_clients = {}
+        # Credit statistics
+        total_credit = 0
+        credit_paid = 0
+        credit_pending = 0
+        # Safe get method for handling potential string or None values
+        def safe_get(obj, key, default=None):
+            if obj is None:
+                return default
+            if isinstance(obj, dict):
+                return obj.get(key, default)
+            return default
+        # Helper function to parse JSON strings
+        def parse_json(json_str, default=None):
+            if not json_str:
+                return default
+            if isinstance(json_str, dict):
+                return json_str
+            try:
+                return json.loads(json_str)
+            except json.JSONDecodeError:
+                return default
+        # Process each patient
+        for patient in patients:
+            # Process total amount
+            try:
+                total_amount = float(safe_get(patient, 'totalAmount', 0))
+                total_revenue += total_amount
+            except (ValueError, TypeError):
+                pass
+            # Extract payment information
+            payment_info = parse_json(safe_get(patient, 'payment_method'), {'paymentmethod': ''})
+            patient_payment_method = safe_get(payment_info, 'paymentmethod', '')
+            # Handle partial payments
+            if patient_payment_method == 'PartialPayment':
+                partial_payment_info = parse_json(safe_get(patient, 'PartialPayment'), {})
+                actual_method = safe_get(partial_payment_info, 'method', '')
+                # Record the partial payment for display in the dashboard
+                payment_methods['PartialPayment'] += 1
+                if actual_method and actual_method in payment_methods:
+                    try:
+                        # Calculate paid amount and credit amount
+                        total_amount = float(safe_get(patient, 'totalAmount', 0))
+                        credit_amount = float(safe_get(partial_payment_info, 'credit', 0))
+                        paid_amount = total_amount - credit_amount
+                        # Add paid amount to the actual payment method
+                        payment_method_amounts[actual_method] += paid_amount
+                        # Add credit amount to the 'Credit' category
+                        payment_method_amounts['Credit'] += credit_amount
+                        # Record total amount under PartialPayment for accurate statistics
+                        payment_method_amounts['PartialPayment'] += total_amount
+                    except (ValueError, TypeError):
+                        pass
+            elif patient_payment_method and patient_payment_method in payment_methods:
+                payment_methods[patient_payment_method] += 1
+                # Add amount to payment method total
+                try:
+                    payment_amount = float(safe_get(patient, 'totalAmount', 0))
+                    payment_method_amounts[patient_payment_method] += payment_amount
+                except (ValueError, TypeError):
+                    pass
+            # Process segment
+            segment = safe_get(patient, 'segment', '')
+            if segment and segment in segments:
+                segments[segment] += 1
+            # Process B2B clients
+            if segment == 'B2B':
+                b2b_name = safe_get(patient, 'B2B', '')
+                if b2b_name:
+                    b2b_clients[b2b_name] = b2b_clients.get(b2b_name, 0) + 1
+            # Process credit information - from both direct credit and partial payments
+            try:
+                # Direct credit amount
+                credit_amount = float(safe_get(patient, 'credit_amount', 0))
+                # Add credit from partial payments if not already included
+                if not credit_amount and patient_payment_method == 'PartialPayment':
+                    partial_payment_info = parse_json(safe_get(patient, 'PartialPayment'), {})
+                    partial_credit = float(safe_get(partial_payment_info, 'credit', 0))
+                    credit_amount += partial_credit
+                total_credit += credit_amount
+            except (ValueError, TypeError):
+                pass
+            # Process credit details for paid amounts
+            credit_details = parse_json(safe_get(patient, 'credit_details'), [])
+            if isinstance(credit_details, list):
+                for detail in credit_details:
+                    if isinstance(detail, dict):
+                        try:
+                            amount_paid = float(safe_get(detail, 'amount_paid', 0))
+                            credit_paid += amount_paid
+                        except (ValueError, TypeError):
+                            pass
+        credit_pending = total_credit - credit_paid
+        # Prepare payment method statistics for the filtered view
+        filtered_payment_stats = {}
+        if payment_method:
+            filtered_payment_stats = {
+                'count': 0,
+                'amount': 0
+            }
+            # Count patients with the specified payment method (including partial payments)
+            for patient in patients:
+                payment_info = parse_json(safe_get(patient, 'payment_method'), {'paymentmethod': ''})
+                patient_payment_method = safe_get(payment_info, 'paymentmethod', '')
+                is_matching = False
+                if patient_payment_method == payment_method:
+                    is_matching = True
+                elif patient_payment_method == 'PartialPayment':
+                    partial_payment_info = parse_json(safe_get(patient, 'PartialPayment'), {})
+                    actual_method = safe_get(partial_payment_info, 'method', '')
+                    if actual_method == payment_method:
+                        is_matching = True
+                if is_matching:
+                    filtered_payment_stats['count'] += 1
+                    try:
+                        filtered_payment_stats['amount'] += float(safe_get(patient, 'totalAmount', 0))
+                    except (ValueError, TypeError):
+                        pass
+        # Prepare response data
+        response_data = {
+            'total_patients': total_patients,
+            'total_revenue': round(total_revenue, 2),
+            'payment_methods': payment_methods,
+            'payment_method_amounts': {k: round(v, 2) for k, v in payment_method_amounts.items()},
+            'segments': segments,
+            'b2b_clients': dict(sorted(b2b_clients.items(), key=lambda x: x[1], reverse=True)),
+            'credit_statistics': {
+                'total_credit': round(total_credit, 2),
+                'credit_paid': round(credit_paid, 2),
+                'credit_pending': round(credit_pending, 2)
+            }
+        }
+        if payment_method:
+            response_data['filtered_payment_stats'] = {
+                'count': filtered_payment_stats['count'],
+                'amount': round(filtered_payment_stats['amount'], 2)
+            }
+        return JsonResponse({
+            'success': True,
+            'data': response_data
+        })
+    except Exception as e:
+        return JsonResponse({
+            'success': False,
+            'error': str(e)
+        }, status=500)
